+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - render HTML tags properly in messages
 - basic support for udiff format in diff viewer
 - excluding information about disabled power tools in the system prompt
+- improved context files messages in Agent
 
 ## [0.22.0]